Move GUI plugin types into their own file
