+++ src/components/scenes/PluginViewLegacyScene.js
@@ -19,7 +19,7 @@ import { PluginBridge, pop as pluginPop } from '../../modules/UI/scenes/Plugins/
 import * as UI_SELECTORS from '../../modules/UI/selectors.js'
 import type { GuiMakeSpendInfo } from '../../reducers/scenes/SendConfirmationReducer.js'
 import styles from '../../styles/scenes/PluginsStyle.js'
-import { type BuySellPlugin, type PluginUrlMap } from '../../types/types.js'
+import { type BuySellPlugin, type PluginUrlMap } from '../../types/GuiPluginTypes.js'
 import { SceneWrapper } from '../common/SceneWrapper.js'
 import { showError } from '../services/AirshipInstance.js'
